Fix DirectMapper unittest
